requirements.txt
streamlit
pandas
plotly
supabase
google-genai
python-dotenv
